modify testbplus param

+++ libbrcdchain/State.h
@@ -479,7 +479,7 @@ public:
     
     void transferAutoEx(std::vector<std::shared_ptr<transationTool::operation>> const& _ops, h256 const& _trxid, int64_t _timeStamp, u256 const& _baseGas);
 
-    void testBplus(std::vector<std::shared_ptr<transationTool::operation>> const& _ops, int32_t const& _time);
+    void testBplus(std::vector<std::shared_ptr<transationTool::operation>> const& _ops, int64_t const& _time);
 
 private:
     void addSysVoteDate(Address const& _sysAddress, Address const& _id);
